Providing the locomotion movement to the player
In this commit I have created a prefab of locomotion movement in a unique project and import the prefab into the repository. This movement is continuous meaning that the movement is similar towards left and right stick controller input with left stick moving and right stick turning.

However there is a possible issue with the prefab that was made with a interaction toolkit in the versions 2, the version used for this repository is a version 3, meaning that I might have to make changes if necessary.