PlayerModel & PlayerController
In this commit, I added the player model and player controller.

In the player model includes a float variable representing health with a GetHealth function that  can grab the variable to another script.

In player Controller there is a function however is on non-edit mode (//) meaning the function won't run unless you remove all // in the void.