Updated Taking Damage
Taking damage void is now on the player controller which will call reducing damage from the player model (TakeDamage void has been changed to ReduceHealth).

In addition the take damage void in the player controller will check if the player's health is below or equal to 0 in which will provide a game over when below 0.