Spawn Manager & Spawn Data
Spawn manager has a Spawn method for depending on the amount of type of enemy will loop towards Instantiating the object until the integer to track the amount of instantiate that occur in the loop.

This also includes a randomPoint method that turns to a value of a Vector3, this randomly chooses the range based on simplifying and finding y in the area of a circle (y = SQRT r^2 - x^2) in which y would be the z axis position this will provide the enemy to spawn at a random range of those value depending of the Spawn Data's SpawnRadius value.

Another method in the spawn manager is Height which is he same as randomPoint except is getting the Height in the SpawnData and this only affect the y axis and not x or z, this will only be for the helicopter since the enemy needs to spawn above ground.

SpawnData provides the amount of enemies that need to spawn along with the object of the enemy, radius to spawn the enemy randomly and height for only the helicopter that will spawn them based around the height of the spawner.