Object Pooling Enemies
The object pooling was able to function when spawning the enemies individually (spawning either a set of tanks/helicopters for one spawner) when using the TestSpawning script to check if they spawn.

However spawning multiple enemies doesn't work as attended, this could be how I implement spawning enemies into the game, or could be the null reference in the factory when pooling objects.